requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.1
altgraph==0.17.4
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
Authlib==1.6.5
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dash==3.2.0
deepseek==1.0.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
fastapi==0.121.1
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.59.1
frozendict==2.4.6
frozenlist==1.7.0
google-adk==1.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.126.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.49.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.3.2
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaraco.context==6.0.1
Jinja2==3.1.6
jiter==0.10.0
Js2Py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
mcp==1.21.0
more-itertools==10.7.0
MouseInfo==0.1.3
multidict==6.6.3
multitasking==0.0.12
narwhals==2.10.1
nest-asyncio==1.6.0
numpy==2.2.6
openai==0.27.8
opencage==3.2.0
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pefile==2023.2.7
phonenumbers==9.0.10
pillow==11.2.1
pipwin==0.5.2
platformdirs==4.5.0
plotly==6.3.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyjsparser==2.7.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.4
retrying==1.4.2
rpds-py==0.28.0
rsa==4.9.1
selenium==4.33.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.10.0
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
sse-starlette==3.0.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wolframalpha==5.0.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
